+++ demisto_sdk/commands/find_dependencies/find_dependencies.py
@@ -16,7 +16,7 @@ MINIMUM_DEPENDENCY_VERSION = LooseVersion('6.0.0')
 COMMON_TYPES_PACK = 'CommonTypes'
 
 
-def parse_for_pack_metadata(dependency_graph: nx.DiGraph, graph_root: str, verbose: int = 1) -> tuple:
+def parse_for_pack_metadata(dependency_graph: nx.DiGraph, graph_root: str, verbose: bool = False) -> tuple:
     """
     Parses calculated dependency graph and returns first and all level parsed dependency.
     Additionally returns list of displayed pack images of all graph levels.
@@ -24,7 +24,7 @@ def parse_for_pack_metadata(dependency_graph: nx.DiGraph, graph_root: str, verbo
     Args:
         dependency_graph (DiGraph): dependency direct graph.
         graph_root (str): graph root pack id.
-        verbose(int): verbosity level - 1-3.
+        verbose(bool): Whether to print the log to the console.
 
     Returns:
         dict: first level dependencies parsed data.
@@ -41,10 +41,8 @@ def parse_for_pack_metadata(dependency_graph: nx.DiGraph, graph_root: str, verbo
 
     all_level_dependencies = [n for n in dependency_graph.nodes if dependency_graph.in_degree(n) > 0]
 
-    if verbose == 3:
+    if verbose:
         click.secho(f'All level dependencies are: {all_level_dependencies}', fg='white')
-    elif verbose == 2:
-        click.secho(f'First level dependencies are: {first_level_dependencies}', fg='white')
 
     return first_level_dependencies, all_level_dependencies
 
@@ -1132,7 +1130,7 @@ class PackDependencies:
 
     @staticmethod
     def find_dependencies(pack_name: str, id_set_path: str = '', exclude_ignored_dependencies: bool = True,
-                          update_pack_metadata: bool = True, silent_mode: bool = False, verbose: int = 0,
+                          update_pack_metadata: bool = True, silent_mode: bool = False, verbose: bool = False,
                           skip_id_set_creation: bool = False) -> dict:
         """
         Main function for dependencies search and pack metadata update.
@@ -1143,7 +1141,7 @@ class PackDependencies:
             exclude_ignored_dependencies (bool): Determines whether to include unsupported dependencies or not.
             update_pack_metadata (bool): Determines whether to update to pack metadata or not.
             silent_mode (bool): Determines whether to echo the dependencies or not.
-            verbose(int): verbosity level - 1-3
+            verbose(bool): Whether to print the log to the console.
             skip_id_set_creation (bool): Whether to skip id_set.json file creation.
 
         Returns:
@@ -1160,7 +1158,7 @@ class PackDependencies:
                 id_set = json.load(id_set_file)
 
         dependency_graph = PackDependencies.build_dependency_graph(
-            pack_id=pack_name, id_set=id_set, verbose=True if verbose == 3 else False,
+            pack_id=pack_name, id_set=id_set, verbose=verbose,
             exclude_ignored_dependencies=exclude_ignored_dependencies)
         first_level_dependencies, _ = parse_for_pack_metadata(dependency_graph, pack_name, verbose)
         if update_pack_metadata: