+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 * Mappers dependencies from incident types and incident fields are now marked as optional.
 * Classifier dependencies from incident types are now marked as optional.
 * Updated **demisto-sdk init** command to no longer create `created` field in pack_metadata file.
-* Using the *verbose* argument in the **find-dependencies** command will now log to the console with verbosity levels.
+* Using the *verbose* argument in the **find-dependencies** command will now log to the console.
 
 # 1.2.16
 * Added allowed ignore errors to the *IDSetValidator*.